调节了一些prompt,引入术语表
