+++ data/glossary.json
@@ -1,0 +1,17 @@
+{
+    "攻击力": "干员的攻击力，决定了对敌人造成的伤害的基础数值。",
+    "生命值": "干员的生命值，决定了干员能承受的伤害量。",
+    "伤害": "干员对敌人造成的伤害，或者敌人对干员造成的伤害，会影响干员或敌人的生命值。任何伤害都分为物理伤害和法术伤害。法术伤害和物理伤害的减伤计算方式不同。",
+    "防御": "干员针对物理攻击防御力，减少受到的伤害，计算公式为：实际受到的伤害=物理攻击伤害-防御力。",
+    "法抗": "即法术抗性",
+    "法术抗性": "干员针对法术攻击的抗性，减少受到的法术伤害，计算公式为：实际受到的伤害=法术攻击伤害*(1-法术抗性/100)。",
+    "术师": "攻击默认造成法术伤害",
+    "医疗": "不会攻击敌人，而是造成等同于攻击伤害的治疗，不计算防御和法术抗性",
+    "先锋": "攻击默认造成物理伤害",
+    "狙击": "攻击默认造成物理伤害",
+    "特种": "攻击默认造成物理伤害",
+    "辅助": "攻击默认造成法术伤害",
+    "重装": "攻击默认造成物理伤害",
+    "近卫": "攻击默认造成物理伤害",
+    "驭械术师":"初始携带1个浮游单元，浮游单元初始造成相当于干员攻击力20%的法术伤害，每次对上次攻击的同一目标进行攻击时该数值增加15，上限110%（即通常情况下需叠加6次才能达到上限）。"
+}
