+++ requirements.txt
@@ -1,5 +1,5 @@
-mcp[cli]~=1.8.1
+mcp[cli]==1.13.0
 fastapi==0.116.1
 uvicorn==0.34.3
-pydantic==2.9.2
-concurrent_log_handler~=0.9.28
+pydantic==2.11.0
+concurrent_log_handler==0.9.28
