+++ src/assets/glossary_data.py
@@ -1,0 +1,10 @@
+import json
+from pathlib import Path
+
+GLOSSARY = {}
+
+with open('data/glossary.json', 'r') as f:
+    GLOSSARY = json.load(f)
+
+__all__ = ["GLOSSARY"]
+
