+++ src/mcp_tools/__init__.py
@@ -1,4 +1,5 @@
 from . import (
     operator_basic,
     operator_skill,
+    arknights_glossary,
 )
