+++ src/server.py
@@ -1,5 +1,8 @@
+import logging
 from fastapi import FastAPI
 from mcp.server.fastmcp import FastMCP
+from functools import wraps
+
 
 # 创建 FastAPI 应用
 app = FastAPI()
@@ -10,8 +13,10 @@ app = FastAPI()
 #   "url": "http://localhost:9000/mcp/sse"
 # }
 mcp = FastMCP("明日方舟知识库")
-mcp.settings.mount_path = "/mcp"
 app.mount("/mcp", mcp.sse_app())
+
+mcp_logger = logging.getLogger("mcp")
+mcp_logger.setLevel(logging.INFO)  # 默认日志等级
 
 # 定义一个简单的状态检查路由
 @app.get("/rest/status")
